+++ between-network/webServer/models/operLog.js
@@ -3,7 +3,7 @@ const Sequelize = require('sequelize')
 module.exports = function(sequelize,DataTypes){
     var User = sequelize.define('oper_log',{
         id:{
-            type:DataTypes.INTEGER,
+            type:DataTypes.BIGINT,
             primaryKey:true,
             allowNull:false,
             autoIncrement:true
@@ -18,10 +18,10 @@ module.exports = function(sequelize,DataTypes){
             type:DataTypes.STRING
         },
         u_id:{
-            type:DataTypes.INTEGER
+            type:DataTypes.BIGINT
         },
         o_id:{
-            type:DataTypes.INTEGER
+            type:DataTypes.BIGINT
         },
         type:{
             type:DataTypes.INTEGER
